+++ src/components/select-field/index.jsx
@@ -1,4 +1,5 @@
 import { FormGroup, FormText, Label, Select } from '@triniti/admin-ui-plugin/components';
+import noop from 'lodash/noop';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
@@ -12,6 +13,7 @@ const SelectField = ({
   label,
   meta: { error },
   multi,
+  onInputChange: handleInputChange,
   ...rest
 }) => {
   const arrowRenderer = rest.creatable ? () => null : undefined;
@@ -22,15 +24,16 @@ const SelectField = ({
       {label && <Label for={input.name}>{label}</Label>}
       <Select
         arrowRenderer={arrowRenderer}
+        classNamePrefix="Select"
+        disabled={disabled}
         multi={multi}
         name={input.name}
-        value={input.value || ''}
-        onChange={input.onChange}
         onBlur={() => input.onBlur(input.value)}
-        disabled={disabled}
-        classNamePrefix="Select"
+        onChange={input.onChange}
+        onInputChange={(text) => handleInputChange(text, input)}
+        value={input.value || ''}
         styles={{
-          menuList: (base) => ({
+          menuList: () => ({
             maxHeight: 'unset',
           }),
         }}
@@ -50,6 +53,7 @@ SelectField.propTypes = {
   label: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
   meta: PropTypes.shape({}).isRequired,
   multi: PropTypes.bool,
+  onInputChange: PropTypes.func,
 };
 
 SelectField.defaultProps = {
@@ -59,6 +63,7 @@ SelectField.defaultProps = {
   hasBorder: false,
   label: '',
   multi: false,
+  onInputChange: noop,
 };
 
 export default SelectField;