+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ## patch
+* allow seo meta keywords and hashtags to be selected using comma key
 * update site-logo to new triniti logo
 * issue #6: Modal does not auto-close when selecting a new block
 * style (in triniti) the save buttons